[Refactoring] Move the components for the header to the base folder
- header에 대한 component들을 base 폴더로 이동
